requirements.txt
requests>=2.28.1
python-dotenv>=0.21.0
urllib3>=1.26.12 